Object storage (#3)
* Improve ETL (#2)

* add prepare-data

* paginate queries

* fixes

* add language + fixes

* replace TRUNCATE by UPSERT strategy

* Local storage

* add  todo

+++ etl/src/movie_etl/sql/ddl/movie.sql
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS movie (
     revenue BIGINT DEFAULT NULL,
     qty_ratings INT NOT NULL DEFAULT 0,
     avg_rating NUMERIC(3, 2),
-    min_rating INT,
-    max_rating INT,
+    min_rating NUMERIC(3, 2),
+    max_rating NUMERIC(3, 2),
 
     UNIQUE (title, release_date)
 );
